fix refactor API
